Modify the boardsSlice with try/catch. Use the boards state in Sidebar and Navbar combo box.
